+++ spec/models/widget_spec.rb
@@ -121,8 +121,8 @@ describe Widget, :type => :model do
 
   describe "Methods" do
     describe "Instance", :versioning => true do
-      describe '#originator' do
-        it { is_expected.to respond_to(:originator) }
+      describe '#paper_trail_originator' do
+        it { is_expected.to respond_to(:paper_trail_originator) }
 
         describe "return value" do
           let(:orig_name) { Faker::Name.name }
@@ -133,9 +133,9 @@ describe Widget, :type => :model do
             specify { expect(widget).to be_live }
 
             it "should return the originator for the model at a given state" do
-              expect(widget.originator).to eq(orig_name)
+              expect(widget.paper_trail_originator).to eq(orig_name)
               widget.whodunnit(new_name) { |w| w.update_attributes(:name => 'Elizabeth') }
-              expect(widget.originator).to eq(new_name)
+              expect(widget.paper_trail_originator).to eq(new_name)
             end
           end
 
@@ -150,7 +150,7 @@ describe Widget, :type => :model do
               let(:reified_widget) { widget.versions[1].reify }
 
               it "should return the appropriate originator" do
-                expect(reified_widget.originator).to eq(orig_name)
+                expect(reified_widget.paper_trail_originator).to eq(orig_name)
               end
 
               it "should not create a new model instance" do
@@ -162,7 +162,7 @@ describe Widget, :type => :model do
               let(:reified_widget) { widget.versions[1].reify(:dup => true) }
 
               it "should return the appropriate originator" do
-                expect(reified_widget.originator).to eq(orig_name)
+                expect(reified_widget.paper_trail_originator).to eq(orig_name)
               end
 
               it "should not create a new model instance" do
@@ -170,6 +170,27 @@ describe Widget, :type => :model do
               end
             end
           end
+        end
+      end
+
+      describe "#originator" do
+        subject { widget }
+
+        it { is_expected.to respond_to(:originator) }
+        let(:warning_msg) do
+          "DEPRECATED: use `paper_trail_originator` instead of `originator`." +
+          " Support for `originator` will be removed in PaperTrail 4.0"
+        end
+
+        it 'should set the invoke `paper_trail_originator`' do
+          is_expected.to receive(:warn)
+          is_expected.to receive(:paper_trail_originator)
+          subject.originator
+        end
+
+        it 'should display a deprecation warning' do
+          is_expected.to receive(:warn).with(warning_msg)
+          subject.originator
         end
       end